Move tests/ to bop/tests
The tests/ director only had bop tests. This follows a more standard directory structure
