+++ bop/tests/add/rakefile
