+++ bop/tests/str/BOP_strcmp/rakefile
