+++ bop/tests/str/BOP_string/rakefile
