+++ bop/tests/str/r_chars.rb
