+++ bop/tests/str/strsub2/rakefile
