+++ bop/tests/unit_tests/ary_bitmap_tests.c
