+++ rakefile
@@ -1,6 +1,5 @@
 task :default => :ruby
-test_dir = ['add', 'sleep'] #TODO enable string tests, 'str/BOP_string', 'str/strsub', 'str/strsub2']
-#TODO add back str/BOP_strcmp
+
 task :ruby_setup do
   cd "ruby" do
     if (!File.exists?("Makefile")) then
@@ -68,17 +67,10 @@ task :ruby_test => :ruby do
 end
 
 task :bop_test => :bop do
-  sh 'make -C ./bop/ clean'
-  test_dir.each do |dir|
-		cd 'tests/' + dir do
-			sh 'rake; rake run' do |ok, res|
-        if ! ok then
-          fail "bop tests failed status: #{res.exitstatus}"
-        end
-	  	end
-    end
+  cd 'bop' do
+    sh 'make --silent -B DEBUG=1'
+    sh 'make test'
   end
-  puts "Finished running all tests"
 end
 
 task :test do
@@ -97,11 +89,6 @@ task :clean do
 	end
   cd "bop" do
     sh 'make clean'
-  end
-  test_dir.each do |dir|
-    cd "tests/" + dir do
-      sh 'rake clean'
-    end
   end
 end
 task :again do
